requirements.txt
azure-core
azure-identity
azure-storage-blob